+++ src/FacebookSharp.SqlProvider/SqlFacebookMembershipProvider.cs
@@ -103,7 +103,7 @@ namespace FacebookSharp
             }
         }
 
-        public void LinkFacebook(string membershipUsername, string facebookId, string accessToken, int expiresIn)
+        public void LinkFacebook(string membershipUsername, string facebookId, string accessToken, long expiresIn)
         {
             // todo: add expires in
             LinkFacebook(membershipUsername, facebookId, accessToken);
@@ -118,7 +118,7 @@ namespace FacebookSharp
             LinkFacebook(user.UserName, facebookId, accessToken);
         }
 
-        public void LinkFacebook(object membershipProviderUserKey, string facebookId, string accessToken, int expiresIn)
+        public void LinkFacebook(object membershipProviderUserKey, string facebookId, string accessToken, long expiresIn)
         {
             // todo: add expires in
             LinkFacebook(membershipProviderUserKey, facebookId, accessToken);
